+++ ecommerce/android/Getflix/app/src/main/java/com/example/getflix/activities/MainActivity.kt
@@ -7,13 +7,13 @@ import android.view.MotionEvent
 import android.view.inputmethod.InputMethodManager
 import androidx.fragment.app.Fragment
 import com.example.getflix.R
-import com.example.getflix.ui.fragment.CartFragment
-import com.example.getflix.ui.fragment.CategoriesFragment
-import com.example.getflix.ui.fragment.HomePageFragment
-import com.example.getflix.ui.fragment.ProfileFragment
+import com.example.getflix.ui.fragments.CartFragment
+import com.example.getflix.ui.fragments.CategoriesFragment
+import com.example.getflix.ui.fragments.HomePageFragment
+import com.example.getflix.ui.fragments.ProfileFragment
 
 
-import com.example.getflix.ui.fragment.*
+import com.example.getflix.ui.fragments.*
 
 
 import kotlinx.android.synthetic.main.activity_main.*
