+++ ecommerce/android/Getflix/app/src/main/java/com/example/getflix/ui/fragments/CartFragment.kt
@@ -1,4 +1,4 @@
-package com.example.getflix.ui.fragment
+package com.example.getflix.ui.fragments
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -7,15 +7,12 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.databinding.DataBindingUtil
 import androidx.lifecycle.ViewModelProvider
-import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.example.getflix.R
 import com.example.getflix.databinding.FragmentCartBinding
 import com.example.getflix.models.ProductModel
 import com.example.getflix.ui.adapters.CartAdapter
-import com.example.getflix.ui.adapters.FavoritesAdapter
 import com.example.getflix.ui.viewmodels.CartViewModel
-import kotlinx.android.synthetic.main.fragment_cart.*
 
 import kotlinx.android.synthetic.main.activity_main.*
 import kotlinx.android.synthetic.main.activity_main.view.*
