+++ ecommerce/android/Getflix/app/src/main/java/com/example/getflix/ui/fragments/HomePageFragment.kt
@@ -1,4 +1,4 @@
-package com.example.getflix.ui.fragment
+package com.example.getflix.ui.fragments
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -11,7 +11,7 @@ import androidx.lifecycle.ViewModelProvider
 import androidx.navigation.fragment.NavHostFragment
 import com.example.getflix.R
 import com.example.getflix.databinding.FragmentHomePageBinding
-import com.example.getflix.viewmodels.HomeViewModel
+import com.example.getflix.ui.viewmodels.HomeViewModel
 import kotlinx.android.synthetic.main.activity_main.*
 import kotlinx.android.synthetic.main.activity_main.view.*
