+++ ecommerce/android/Getflix/app/src/main/java/com/example/getflix/ui/viewmodels/HomeViewModel.kt
@@ -1,4 +1,4 @@
-package com.example.getflix.viewmodels
+package com.example.getflix.ui.viewmodels
 
 import android.app.Application
 import androidx.lifecycle.*
