+++ ecommerce/android/Getflix/app/src/main/res/layout/fragment_favorites.xml
@@ -12,7 +12,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.fragment.LoginFragment"
+        tools:context=".ui.fragments.LoginFragment"
         android:layout_marginTop="?attr/actionBarSize">
