+++ ecommerce/android/Getflix/app/src/main/res/layout/fragment_login.xml
@@ -12,7 +12,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background="#52EBD494"
-        tools:context=".ui.fragment.LoginFragment">
+        tools:context=".ui.fragments.LoginFragment">
 
         <TextView
             android:id="@+id/textView"
